refactor: trigger edit mode 和 view mode 合併
