+++ dashboard/src/index.scss
@@ -23,7 +23,6 @@
 @import './pages/oauth-client/oauth-client.scss';
 @import './pages/triggers/triggers.scss';
 @import './pages/trigger/trigger.scss';
-@import './pages/trigger/trigger-form/trigger-form.scss';
 @import './pages/not-found/not-found.scss';
 
 @import './components/footer/footer.scss';
