--- dashboard/src/pages/trigger/trigger-form/trigger-form.scss
@@ -1,11 +1,0 @@
-.trigger-form-wrapper {
-  width: 720px;
-
-  @include media-breakpoint-down(lg) {
-    width: 100%;
-  }
-
-  .title {
-    cursor: pointer;
-  }
-}
