+++ dashboard/src/pages/trigger/trigger.scss
@@ -1,2 +1,11 @@
 .trigger {
+    width: 720px;
+
+    @include media-breakpoint-down(lg) {
+        width: 100%;
+    }
+
+    .title {
+        cursor: pointer;
+    }
 }
